Add Webpack to automatically watch and build changes.
