+++ config/paths.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const path = require('path');
+
+const PATHS = {
+  src: path.resolve(__dirname, '../src'),
+  build: path.resolve(__dirname, '../build'),
+};
+
+module.exports = PATHS;
