+++ public/api/config.sample.js
@@ -1,0 +1,8 @@
+/* config.js */
+
+var CONFIG = {
+    api_key: '',
+    api_secret: '',
+    api_url: 'https://ws.audioscrobbler.com/2.0/',
+    cb_file: 'cb.html',
+};
