+++ public/api/lastfm.js
