+++ public/popup.html
@@ -1,0 +1,102 @@
+<!doctype html>
+<html>
+
+<head>
+  <title>NhacCuaTui Scrobbler</title>
+  <link rel="stylesheet" href="https://unpkg.com/purecss@0.6.2/build/pure-min.css"
+    integrity="sha384-UQiGfs9ICog+LwheBSRCt1o5cbyKIHbwjWscjemyBMT9YCUMZffs6UqUTd0hObXD" crossorigin="anonymous">
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
+    crossorigin="anonymous">
+  <style>
+    body {
+      color: #333;
+      width: 240px;
+    }
+
+    a {
+      color: #2582be;
+    }
+
+    #heading {
+      background: #111;
+      background-image: linear-gradient(0deg, #000, #333);
+      color: #eee;
+      display: inline-block;
+      text-align: center;
+      width: 100%;
+    }
+
+    #heading h3 {
+      margin: 8px;
+    }
+
+    #container {
+      background: #fff;
+      background-image: linear-gradient(0deg, #fff, #fff, #f1f1f1);
+      padding: 8px;
+    }
+
+    #status {
+      text-align: right;
+    }
+
+    #notPlaying {
+      display: none;
+    }
+
+    #nowPlaying {
+      display: block;
+      margin-bottom: 16px;
+      min-height: 48px;
+    }
+
+    #scrobbled {
+      color: #2582be;
+      display: none;
+    }
+  </style>
+  <script type="text/javascript" src="vendor/jquery-3.2.1.min.js"></script>
+  <script type="text/javascript" src="popup.js"></script>
+</head>
+
+<body>
+  <div id="heading">
+    <h3>NhacCuaTui Scrobbler</h3>
+  </div>
+  <div id="container">
+    <div class="pure-g">
+      <div class="pure-u-5-5">
+        <div id="nowPlaying">
+          <div class="pure-g">
+            <div class="pure-u-1-5">
+              <i class="fa fa-headphones fa-3x"></i>
+            </div>
+            <div class="pure-u-4-5">
+              <strong><span id="track">Nothing is playing!</span></strong></br>
+              <span id="artist"></span></br>
+              <span id="scrobbled">Scrobbled</span>
+            </div>
+          </div>
+        </div>
+        <div id="status">
+          <div id="loggedIn">
+            <a id="lastFMButton" href="#" target="_blank" class="pure-button pure-button-primary">
+              <i class="fa fa-lastfm"></i> <span id="userName"></span>
+            </a>
+            <button id="logOutButton" class="pure-button pure-button">
+              <i class="fa fa-sign-out"></i> Log out
+            </button>
+          </div>
+          <div id="loggedOut">
+            <button id="logInButton" class="pure-button pure-button-primary">
+              <i class="fa fa-sign-in"></i> Log in
+            </button>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+  </div>
+</body>
+
+</html>
